requirements.txt
requests==2.32.5
flask==3.1.2
apscheduler==3.11.0
python-dotenv==1.1.1
